+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version 0.23
 - add support for changing column `lco` with values present for many cases
 - removed implicit from datalink to jsonb cast
 - lco number assignments in `link_control_options` again to be more logical
+- some boolean args changes to integer to be more consistent
 
 Version 0.22
 ------------